parser: Change generic and port list syntax.
Reasons: readability, parseability.

Reserve <> and () delimiters for the positional generic and
port assignments.
Those will use commas instead of semicolons as separators.

In the future, short statements like

    put Csmall<100e-9>(vdd, gnd);

equilavent to

    put Csmall
    generic {
        capacitance := 100e-9;
    }
    port {
        pos == vdd;
        neg == gnd;
    };

will be supported.

Switch to more VHDL-like "generic { }" and "port { }" for both
declarations and named assignments.

Disambiguating named and positional assignments within
the same kind of block wouldn't be easy when arbitrary wire and
generic expressions are allowed in the future.